backend/api/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from .models import Schema, Product

class SchemaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Schema
        fields = '__all__'

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'
